+++ packages/image_picker/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerDelegate.java
@@ -12,6 +12,7 @@ import android.content.pm.ResolveInfo;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.provider.MediaStore;
+import android.support.annotation.VisibleForTesting;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.FileProvider;
 import io.flutter.plugin.common.MethodCall;
@@ -64,27 +65,119 @@ import java.util.UUID;
 public class ImagePickerDelegate
     implements PluginRegistry.ActivityResultListener,
         PluginRegistry.RequestPermissionsResultListener {
-  private static final int REQUEST_CODE_CHOOSE_FROM_GALLERY = 2342;
-  private static final int REQUEST_CODE_TAKE_WITH_CAMERA = 2343;
+  @VisibleForTesting static final int REQUEST_CODE_CHOOSE_FROM_GALLERY = 2342;
+  @VisibleForTesting static final int REQUEST_CODE_TAKE_WITH_CAMERA = 2343;
+  @VisibleForTesting static final int REQUEST_EXTERNAL_STORAGE_PERMISSION = 2344;
+  @VisibleForTesting static final int REQUEST_CAMERA_PERMISSION = 2345;
 
-  private static final int REQUEST_EXTERNAL_STORAGE_PERMISSION = 2344;
-  private static final int REQUEST_CAMERA_PERMISSION = 2345;
+  @VisibleForTesting final String fileProviderName;
 
   private final Activity activity;
   private final File externalFilesDirectory;
   private final ImageResizer imageResizer;
-  private final String providerName;
+  private final PermissionManager permissionManager;
+  private final IntentResolver intentResolver;
+  private final FileUriResolver fileUriResolver;
+  private final FileUtils fileUtils;
+
+  interface PermissionManager {
+    boolean isPermissionGranted(String permissionName);
+
+    void askForPermission(String permissionName, int requestCode);
+  }
+
+  interface IntentResolver {
+    boolean resolveActivity(Intent intent);
+  }
+
+  interface FileUriResolver {
+    Uri resolveFileProviderUriForFile(String fileProviderName, File imageFile);
+
+    void getFullImagePath(Uri imageUri, OnPathReadyListener listener);
+  }
+
+  interface OnPathReadyListener {
+    void onPathReady(String path);
+  }
 
   private Uri pendingCameraImageUri;
   private MethodChannel.Result pendingResult;
   private MethodCall methodCall;
 
   public ImagePickerDelegate(
-      Activity activity, File externalFilesDirectory, ImageResizer imageResizer) {
+      final Activity activity, File externalFilesDirectory, ImageResizer imageResizer) {
+    this(
+        activity,
+        externalFilesDirectory,
+        imageResizer,
+        null,
+        null,
+        new PermissionManager() {
+          @Override
+          public boolean isPermissionGranted(String permissionName) {
+            return ActivityCompat.checkSelfPermission(activity, permissionName)
+                == PackageManager.PERMISSION_GRANTED;
+          }
+
+          @Override
+          public void askForPermission(String permissionName, int requestCode) {
+            ActivityCompat.requestPermissions(activity, new String[] {permissionName}, requestCode);
+          }
+        },
+        new IntentResolver() {
+          @Override
+          public boolean resolveActivity(Intent intent) {
+            return intent.resolveActivity(activity.getPackageManager()) != null;
+          }
+        },
+        new FileUriResolver() {
+          @Override
+          public Uri resolveFileProviderUriForFile(String fileProviderName, File file) {
+            return FileProvider.getUriForFile(activity, fileProviderName, file);
+          }
+
+          @Override
+          public void getFullImagePath(final Uri imageUri, final OnPathReadyListener listener) {
+            MediaScannerConnection.scanFile(
+                activity,
+                new String[] {imageUri.getPath()},
+                null,
+                new MediaScannerConnection.OnScanCompletedListener() {
+                  @Override
+                  public void onScanCompleted(String path, Uri uri) {
+                    listener.onPathReady(path);
+                  }
+                });
+          }
+        },
+        new FileUtils());
+  }
+
+  /**
+   * This constructor is used exclusively for testing; it can be used to provide mocks to final
+   * fields of this class. Otherwise those fields would have to be mutable and visible.
+   */
+  @VisibleForTesting
+  ImagePickerDelegate(
+      Activity activity,
+      File externalFilesDirectory,
+      ImageResizer imageResizer,
+      MethodChannel.Result result,
+      MethodCall methodCall,
+      PermissionManager permissionManager,
+      IntentResolver intentResolver,
+      FileUriResolver fileUriResolver,
+      FileUtils fileUtils) {
     this.activity = activity;
     this.externalFilesDirectory = externalFilesDirectory;
     this.imageResizer = imageResizer;
-    this.providerName = activity.getPackageName() + ".flutter.image_provider";
+    this.fileProviderName = activity.getPackageName() + ".flutter.image_provider";
+    this.pendingResult = result;
+    this.methodCall = methodCall;
+    this.permissionManager = permissionManager;
+    this.intentResolver = intentResolver;
+    this.fileUriResolver = fileUriResolver;
+    this.fileUtils = fileUtils;
   }
 
   public void chooseImageFromGallery(MethodCall methodCall, MethodChannel.Result result) {
@@ -93,23 +186,13 @@ public class ImagePickerDelegate
       return;
     }
 
-    boolean hasExternalStoragePermission =
-        ActivityCompat.checkSelfPermission(activity, Manifest.permission.READ_EXTERNAL_STORAGE)
-            == PackageManager.PERMISSION_GRANTED;
-
-    if (!hasExternalStoragePermission) {
-      requestExternalStoragePermission();
+    if (!permissionManager.isPermissionGranted(Manifest.permission.READ_EXTERNAL_STORAGE)) {
+      permissionManager.askForPermission(
+          Manifest.permission.READ_EXTERNAL_STORAGE, REQUEST_EXTERNAL_STORAGE_PERMISSION);
       return;
     }
 
     launchPickImageFromGalleryIntent();
-  }
-
-  private void requestExternalStoragePermission() {
-    ActivityCompat.requestPermissions(
-        activity,
-        new String[] {Manifest.permission.READ_EXTERNAL_STORAGE},
-        REQUEST_EXTERNAL_STORAGE_PERMISSION);
   }
 
   private void launchPickImageFromGalleryIntent() {
@@ -125,26 +208,17 @@ public class ImagePickerDelegate
       return;
     }
 
-    boolean hasCameraPermission =
-        ActivityCompat.checkSelfPermission(activity, Manifest.permission.CAMERA)
-            == PackageManager.PERMISSION_GRANTED;
-
-    if (!hasCameraPermission) {
-      requestCameraPermission();
+    if (!permissionManager.isPermissionGranted(Manifest.permission.CAMERA)) {
+      permissionManager.askForPermission(Manifest.permission.CAMERA, REQUEST_CAMERA_PERMISSION);
       return;
     }
 
     launchTakeImageWithCameraIntent();
   }
 
-  private void requestCameraPermission() {
-    ActivityCompat.requestPermissions(
-        activity, new String[] {Manifest.permission.CAMERA}, REQUEST_CAMERA_PERMISSION);
-  }
-
   private void launchTakeImageWithCameraIntent() {
     Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-    boolean canTakePhotos = intent.resolveActivity(activity.getPackageManager()) != null;
+    boolean canTakePhotos = intentResolver.resolveActivity(intent);
 
     if (!canTakePhotos) {
       finishWithError("no_available_camera", "No cameras available for taking pictures.");
@@ -154,7 +228,7 @@ public class ImagePickerDelegate
     File imageFile = createTemporaryWritableImageFile();
     pendingCameraImageUri = Uri.parse("file:" + imageFile.getAbsolutePath());
 
-    Uri imageUri = FileProvider.getUriForFile(activity, providerName, imageFile);
+    Uri imageUri = fileUriResolver.resolveFileProviderUriForFile(fileProviderName, imageFile);
     intent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri);
     grantUriPermissions(intent, imageUri);
 
@@ -226,7 +300,7 @@ public class ImagePickerDelegate
 
   private void handleChoosePictureResult(int resultCode, Intent data) {
     if (resultCode == Activity.RESULT_OK && data != null) {
-      String path = FileUtils.getPathFromUri(activity, data.getData());
+      String path = fileUtils.getPathFromUri(activity, data.getData());
       handleResult(path);
       return;
     }
@@ -237,13 +311,11 @@ public class ImagePickerDelegate
 
   private void handleTakePictureResult(int resultCode) {
     if (resultCode == Activity.RESULT_OK) {
-      MediaScannerConnection.scanFile(
-          activity,
-          new String[] {pendingCameraImageUri.getPath()},
-          null,
-          new MediaScannerConnection.OnScanCompletedListener() {
+      fileUriResolver.getFullImagePath(
+          pendingCameraImageUri,
+          new OnPathReadyListener() {
             @Override
-            public void onScanCompleted(String path, Uri uri) {
+            public void onPathReady(String path) {
               handleResult(path);
             }
           });